+++ src/ResourceManager/DataFactoryV2/Commands.DataFactoryV2/Triggers/StopAzureDataFactoryTriggerCommand.cs
@@ -22,7 +22,7 @@ using Microsoft.Azure.Commands.DataFactoryV2.Properties;
 namespace Microsoft.Azure.Commands.DataFactoryV2
 {
     [Cmdlet(VerbsLifecycle.Stop, Constants.Trigger, DefaultParameterSetName = ParameterSetNames.ByFactoryName, SupportsShouldProcess = true),
-        OutputType(typeof(List<PSTrigger>), typeof(PSTrigger))]
+        OutputType(typeof(PSTrigger))]
     public class StopAzureDataFactoryTriggerCommand : DataFactoryContextActionBaseCmdlet
     {
         [Parameter(ParameterSetName = ParameterSetNames.ByFactoryName, Position = 2, Mandatory = true, ValueFromPipelineByPropertyName = true,