+++ src/ResourceManager/DataLakeAnalytics/Commands.DataLakeAnalytics/ChangeLog.md
@@ -19,6 +19,13 @@
 -->
 ## Current Release
 
+## Version 5.1.0
+* Add support for Catalog ACLs through the following commands:
+    - Get-AzureRmDataLakeAnalyticsCatalogItemAclEntry
+    - Set-AzureRmDataLakeAnalyticsCatalogItemAclEntry
+    - Remove-AzureRmDataLakeAnalyticsCatalogItemAclEntry
+* Fixed formatting of OutputType in help files
+
 ## Version 5.0.0
 * Set minimum dependency of module to PowerShell 5.0
 * Remove deprecated `Tags` alias from cmdlets